Refactored code to remove redundant files, added chc classes with it's vmt equivalent and renamed rules/vmt to rules/transpilation consisting both vmt and chc.
