+++ test/tla/ArithOper.tla
@@ -2,20 +2,16 @@
 EXTENDS Integers
 
 VARIABLE
-    \* @type: Bool;
-    val1,
     \* @type: Int;
     val2,
     \* @type: Int;
     val3
 
 Init ==
-    /\ val1 = FALSE
     /\ val2 = 2
     /\ val3 = 0
 
 m1 ==
-    /\ val1' = ~val1
     /\ val2' = val2 + (1 * 1)
     /\ val2 = (4 - 2)
     /\ UNCHANGED <<val3>>
@@ -23,12 +19,12 @@ m1 ==
 m2 ==
     /\ val3 > 0
     /\ val3' = (100 \div (-3))
-    /\ UNCHANGED <<val1, val2>>
+    /\ UNCHANGED <<val2>>
 
 m3 ==
     /\ val3 <= (99 + 1)
     /\ val3' = val3 + (3 - 2)
-    /\ UNCHANGED <<val1, val2>>
+    /\ UNCHANGED <<val2>>
 
 Next ==
     \/ m1
@@ -39,6 +35,6 @@ Next ==
     \/ m3
 
 inv1 ==
-    /\ val1 = FALSE
+    /\ val3 > 20
     /\ val2 < 100
 =================================================================
