+++ test/tla/ArithOperWithInvSupport.tla
@@ -1,21 +1,17 @@
----------------------------- MODULE ArithOper2 --------------------------
+---------------------------- MODULE ArithOperWithInvSupport --------------------------
 EXTENDS Integers
 
 VARIABLE
-    \* @type: Bool;
-    val1,
     \* @type: Int;
     val2,
     \* @type: Int;
     val3
 
 Init ==
-    /\ val1 = FALSE
     /\ val2 = 2
     /\ val3 = 0
 
 m1 ==
-    /\ val1 = FALSE
     /\ val2' = val2 + 1
     /\ UNCHANGED <<val1, val3>>
