--- test/tla/to-override-dir/run.txt
@@ -1,1 +1,0 @@
- check --temporal=Liveness LongPrefix.tla
