+++ tla-bmcmt/src/test/scala/at/forsyte/apalache/tla/bmcmt/rules/transpilation/TestQuantRule.scala
@@ -1,10 +1,9 @@
-package at.forsyte.apalache.tla.bmcmt.rules.vmt
+package at.forsyte.apalache.tla.bmcmt.rules.transpilation
 
-import at.forsyte.apalache.tla.lir.{BoolT1, ConstT1, FunT1, IntT1, SetT1, TlaEx, TlaType1, TupT1}
+import at.forsyte.apalache.tla.bmcmt.RewriterException
+import at.forsyte.apalache.tla.lir._
 import at.forsyte.apalache.tla.lir.formulas.Booleans._
-import at.forsyte.apalache.tla.lir.formulas.EUF.{Apply, DefineFun, Equal, FunctionVar, ITE}
 import at.forsyte.apalache.tla.lir.formulas._
-import at.forsyte.apalache.tla.pp.UniqueNameGenerator
 import at.forsyte.apalache.tla.typecomp.TBuilderInstruction
 import at.forsyte.apalache.tla.types.tla
 import org.junit.runner.RunWith
@@ -12,7 +11,7 @@ import org.scalatest.funsuite.AnyFunSuite
 import org.scalatestplus.junit.JUnitRunner
 
 @RunWith(classOf[JUnitRunner])
-class TestEUFRule extends AnyFunSuite {
+class TestQuantRule extends AnyFunSuite {
 
   val sType: TlaType1 = ConstT1("SSORT")
   val sSort: UninterpretedSort = UninterpretedSort("SSORT")
@@ -21,14 +20,7 @@ class TestEUFRule extends AnyFunSuite {
 
   val rewriter: ToTermRewriter = ToTermRewriterImpl(constSets)
 
-  val funName: String = "f"
-  val constGen: UniqueNameGenerator = new UniqueNameGenerator {
-    override def newName(): String = funName
-  }
-  val fType: TlaType1 = FunT1(TupT1(sType, IntT1), sType)
-  val f: TBuilderInstruction = tla.name(funName, fType)
-
-  val rule: FormulaRule = new EUFRule(rewriter, new RestrictedSetJudgement(constSets), constGen)
+  val rule: FormulaRule = new QuantifierRule(rewriter, new RestrictedSetJudgement(constSets))
 
   val b: TlaType1 = BoolT1
 
@@ -38,25 +30,23 @@ class TestEUFRule extends AnyFunSuite {
   val qVar: Term = BoolVar("q")
 
   val x: TBuilderInstruction = tla.name("x", sType)
-  val xVar: Term = mkVariable("x", sSort)
-  val xPrimeVar: Term = mkVariable(VMTprimeName("x"), sSort)
   val y: TBuilderInstruction = tla.name("y", IntT1)
   val set: TBuilderInstruction = tla.name("S", SetT1(sType))
   val intSet: TBuilderInstruction = tla.intSet()
 
-  val expected: Map[TlaEx, Term] = Map(
-      tla.assign(tla.prime(x), x).build -> Equal(xPrimeVar, xVar),
-      tla.eql(x, x).build -> Equal(xVar, xVar),
-      tla.ite(p, p, q).build -> ITE(pVar, pVar, qVar),
-      tla.funDef(x, x -> set, y -> intSet).build ->
-        DefineFun(funName, List(("x", sSort), ("y", IntSort)), xVar).asVar,
-      tla.app(f, tla.tuple(x, y)).build ->
-        Apply(FunctionVar(funName, FunctionSort(sSort, sSort, IntSort)), xVar, mkVariable("y", IntSort)),
+  val expected: Map[TlaEx, BoolExpr] = Map(
+      tla.exists(x, set, p).build -> Exists(List(("x", sSort)), pVar),
+      tla.forall(y, intSet, q).build -> Forall(List(("y", IntSort)), qVar),
   )
 
-  test("EUFRule applicability") {
+  test("QuantRule applicability") {
     expected.keys.foreach { ex =>
       assert(rule.isApplicable(ex))
+    }
+
+    assertThrows[RewriterException] {
+      val tType = ConstT1("TSORT")
+      rule(tla.exists(tla.name("t", tType), tla.name("T", SetT1(tType)), p))
     }
 
     val notApp = List(
@@ -73,7 +63,7 @@ class TestEUFRule extends AnyFunSuite {
     }
   }
 
-  test("EUFRule correctness") {
+  test("QuantRule correctness") {
     expected.foreach { case (k, expected) =>
       val actual = rule(k).run(SmtDeclarations.init)._2
       assert(actual == expected)
